Moved helper function in service.ts file to separate util file
